Fix https://bz.apache.org/bugzilla/show_bug.cgi?id=52028 Add support for automatic binding to a free port by a connector if the special value of zero is used for the port. This is mainly useful in embedded and testing scenarios.
This includes backport of the following revisions from Tomcat 7:
r1207695 r1208115 r1208148 r1723441 r1723489

Handling of port number 0 in Jk AJP connector (ChannelSocket, ChannelNioSocket) was changed.
Old behaviour was to disable the connector.
New behaviour is to bind to a random port number.
To disable the connector one can now use a negative port number. I doubt that this "disable" feature was actually used by anyone. Also binding to a random port is not so much different from disabling.

git-svn-id: https://svn.apache.org/repos/asf/tomcat/tc6.0.x/trunk@1723551 13f79535-47bb-0310-9956-ffa450edef68